# api/server/v1/server.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "server/v1/server.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/server/settings": {
      "get": {
        "summary": "Get settings",
        "description": "Returns current PMM Server settings.",
        "operationId": "GetSettings",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetSettingsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "ServerService"
        ]
      },
      "put": {
        "summary": "Change settings",
        "description": "Changes PMM Server settings.",
        "operationId": "ChangeSettings",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ChangeSettingsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1ChangeSettingsRequest"
            }
          }
        ],
        "tags": [
          "ServerService"
        ]
      }
    },
    "/v1/server/settings/readonly": {
      "get": {
        "summary": "Get read-only settings",
        "description": "Returns a stripped version of PMM Server settings.",
        "operationId": "GetReadOnlySettings",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetReadOnlySettingsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "ServerService"
        ]
      }
    },
    "/v1/server/updates": {
      "get": {
        "summary": "Check updates",
        "description": "Checks for available PMM Server updates.",
        "operationId": "CheckUpdates",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CheckUpdatesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "force",
            "description": "If false, cached information may be returned.",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "only_installed_version",
            "description": "If true, only installed version will be in response.",
            "in": "query",
            "required": false,
            "type": "boolean"
          }
        ],
        "tags": [
          "ServerService"
        ]
      }
    },
...
    "/v1/server/updates:getStatus": {
      "post": {
        "summary": "Update status",
        "description": "Returns PMM Server update status.",
        "operationId": "UpdateStatus",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateStatusResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1UpdateStatusRequest"
            }
          }
        ],
        "tags": [
          "ServerService"
        ]
      }
    },
    "/v1/server/updates:start": {
      "post": {
        "summary": "Start update",
        "description": "Starts PMM Server update.",
        "operationId": "StartUpdate",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1StartUpdateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1StartUpdateRequest"
            }
          }
        ],
        "tags": [
          "ServerService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "commonStringArray": {
      "type": "object",
      "properties": {
        "values": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "description": "A wrapper for a string array. This type allows to distinguish between an empty array and a null value."
    },
...
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "v1AdvisorRunIntervals": {
      "type": "object",
      "properties": {
        "standard_interval": {
          "type": "string",
          "description": "Standard check interval."
        },
        "rare_interval": {
          "type": "string",
          "description": "Interval for rare check runs."
        },
        "frequent_interval": {
          "type": "string",
          "description": "Interval for frequent check runs."
        }
      },
      "description": "AdvisorRunIntervals represents intervals between each run of Advisor checks."
    },
...
    "v1ChangeSettingsRequest": {
      "type": "object",
      "properties": {
        "enable_updates": {
          "type": "boolean",
          "x-nullable": true
        },
        "enable_telemetry": {
          "type": "boolean",
          "x-nullable": true
        },
        "metrics_resolutions": {
          "$ref": "#/definitions/v1MetricsResolutions"
        },
        "data_retention": {
          "type": "string",
          "description": "A number of full days for Prometheus and QAN data retention. Should have a suffix in JSON: 2592000s, 43200m, 720h."
        },
        "ssh_key": {
          "type": "string",
          "x-nullable": true
        },
        "aws_partitions": {
          "$ref": "#/definitions/commonStringArray",
          "x-nullable": true
        },
        "enable_advisor": {
          "type": "boolean",
          "x-nullable": true,
          "description": "Enable Advisor."
        },
        "enable_alerting": {
          "type": "boolean",
          "x-nullable": true,
          "description": "Enable Alerting."
        },
        "pmm_public_address": {
          "type": "string",
          "x-nullable": true,
          "description": "PMM Server public address."
        },
        "advisor_run_intervals": {
          "$ref": "#/definitions/v1AdvisorRunIntervals",
          "description": "Intervals between Advisor runs."
        },
        "enable_azurediscover": {
          "type": "boolean",
          "x-nullable": true,
          "description": "Enable Azure Discover."
        },
        "enable_backup_management": {
          "type": "boolean",
          "x-nullable": true,
          "description": "Enable Backup Management."
        },
        "enable_access_control": {
          "type": "boolean",
          "x-nullable": true,
          "title": "Enable Access Control"
        },
        "enable_internal_pg_qan": {
          "type": "boolean",
          "x-nullable": true,
          "description": "Enable Query Analytics for PMM's internal PG database."
        },
        "update_snooze_duration": {
          "type": "string",
          "description": "A number of full days for which an update is snoozed, i.e. a multiple of 24h: 2592000s, 43200m, 720h."
        }
      }
    },
    "v1ChangeSettingsResponse": {
      "type": "object",
      "properties": {
        "settings": {
          "$ref": "#/definitions/v1Settings"
        }
      }
    },
    "v1CheckUpdatesResponse": {
      "type": "object",
      "properties": {
        "installed": {
          "$ref": "#/definitions/v1VersionInfo",
          "description": "Currently installed PMM Server version."
        },
        "latest": {
          "$ref": "#/definitions/v1DockerVersionInfo",
          "description": "Latest available PMM Server version."
        },
        "update_available": {
          "type": "boolean",
          "description": "True if there is a PMM Server update available."
        },
        "latest_news_url": {
          "type": "string",
          "description": "Latest available PMM Server release announcement URL."
        },
        "last_check": {
          "type": "string",
          "format": "date-time",
          "description": "Last check time."
        }
      }
    },
...
    "v1GetReadOnlySettingsResponse": {
      "type": "object",
      "properties": {
        "settings": {
          "$ref": "#/definitions/v1ReadOnlySettings"
        }
      }
    },
    "v1GetSettingsResponse": {
      "type": "object",
      "properties": {
        "settings": {
          "$ref": "#/definitions/v1Settings"
        }
      }
    },
...
    "v1MetricsResolutions": {
      "type": "object",
      "properties": {
        "hr": {
          "type": "string",
          "description": "High resolution. Should have a suffix in JSON: 1s, 1m, 1h."
        },
        "mr": {
          "type": "string",
          "description": "Medium resolution. Should have a suffix in JSON: 1s, 1m, 1h."
        },
        "lr": {
          "type": "string",
          "description": "Low resolution. Should have a suffix in JSON: 1s, 1m, 1h."
        }
      },
      "description": "MetricsResolutions represents Prometheus exporters metrics resolutions."
    },
    "v1ReadOnlySettings": {
      "type": "object",
      "properties": {
        "updates_enabled": {
          "type": "boolean",
          "description": "True if updates are enabled."
        },
        "telemetry_enabled": {
          "type": "boolean",
          "description": "True if telemetry is enabled."
        },
        "advisor_enabled": {
          "type": "boolean",
          "description": "True if Advisor is enabled."
        },
        "alerting_enabled": {
          "type": "boolean",
          "description": "True if Alerting is enabled."
        },
        "pmm_public_address": {
          "type": "string",
          "description": "PMM Server public address."
        },
        "backup_management_enabled": {
          "type": "boolean",
          "description": "True if Backup Management is enabled."
        },
        "azurediscover_enabled": {
          "type": "boolean",
          "description": "True if Azure Discover is enabled."
        },
        "enable_access_control": {
          "type": "boolean",
          "description": "True if Access Control is enabled."
        }
      },
      "description": "ReadOnlySettings represents a stripped-down version of PMM Server settings that can be accessed by users of all roles."
    },
...
    "v1Settings": {
      "type": "object",
      "properties": {
        "updates_enabled": {
          "type": "boolean",
          "description": "True if updates are enabled."
        },
        "telemetry_enabled": {
          "type": "boolean",
          "description": "True if telemetry is enabled."
        },
        "metrics_resolutions": {
          "$ref": "#/definitions/v1MetricsResolutions"
        },
        "data_retention": {
          "type": "string"
        },
        "ssh_key": {
          "type": "string"
        },
        "aws_partitions": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "advisor_enabled": {
          "type": "boolean",
          "description": "True if Advisor is enabled."
        },
        "platform_email": {
          "type": "string"
        },
        "alerting_enabled": {
          "type": "boolean",
          "description": "True if Alerting is enabled."
        },
        "pmm_public_address": {
          "type": "string",
          "description": "PMM Server public address."
        },
        "advisor_run_intervals": {
          "$ref": "#/definitions/v1AdvisorRunIntervals",
          "description": "Intervals between Advisor runs."
        },
        "backup_management_enabled": {
          "type": "boolean",
          "description": "True if Backup Management is enabled."
        },
        "azurediscover_enabled": {
          "type": "boolean",
          "description": "True if Azure Discover is enabled."
        },
        "connected_to_platform": {
          "type": "boolean",
          "title": "True if the PMM instance is connected to Platform"
        },
        "telemetry_summaries": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "Includes list of collected telemetry"
        },
        "enable_access_control": {
          "type": "boolean",
          "description": "True if Access Control is enabled."
        },
        "default_role_id": {
          "type": "integer",
          "format": "int64",
          "description": "Default Access Control role ID for new users."
        },
        "enable_internal_pg_qan": {
          "type": "boolean",
          "description": "True if Query Analytics for PMM's internal PG database is enabled."
        },
        "update_snooze_duration": {
          "type": "string",
          "title": "Duration for which an update is snoozed"
        }
      },
      "description": "Settings represents PMM Server settings."
    },
    "v1StartUpdateRequest": {
      "type": "object",
      "properties": {
        "new_image": {
          "type": "string"
        }
      }
    },
    "v1StartUpdateResponse": {
      "type": "object",
      "properties": {
        "auth_token": {
          "type": "string",
          "description": "Authentication token for getting update statuses."
        },
        "log_offset": {
          "type": "integer",
          "format": "int64",
          "description": "Progress log offset."
        }
      }
    },
    "v1UpdateStatusRequest": {
      "type": "object",
      "properties": {
        "auth_token": {
          "type": "string",
          "description": "Authentication token."
        },
        "log_offset": {
          "type": "integer",
          "format": "int64",
          "description": "Progress log offset."
        }
      }
    },
    "v1UpdateStatusResponse": {
      "type": "object",
      "properties": {
        "log_lines": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "description": "Progress log lines."
        },
        "log_offset": {
          "type": "integer",
          "format": "int64",
          "description": "Progress log offset for the next request."
        },
        "done": {
          "type": "boolean",
          "description": "True when update is done."
        }
      }
    },
    "v1VersionInfo": {
      "type": "object",
      "properties": {
        "version": {
          "type": "string",
          "description": "User-visible version."
        },
        "full_version": {
          "type": "string",
          "description": "Full version for debugging."
        },
        "timestamp": {
          "type": "string",
          "format": "date-time",
          "description": "Build or release date."
        }
      },
      "description": "VersionInfo describes component version, or PMM Server as a whole."
    },
...
  }
}